[yarpdataplayer] made changes for the repeat check not being updated when a new dataset is loaded in the same instance of the yarpdataplayer

+++ src/yarpdataplayer-qt/src/worker.cpp
@@ -290,9 +290,7 @@ void MasterThread::runNormally()
             }
         }
     }
-    this->setRate( (2 / utilities->speed) );
-
-    //fprintf(stdout, "rate is %lf \n", (2 / utilities->speed));
+    this->setRate( (int) (2 / utilities->speed) );
     
     virtualTime += 0.002; // increase by two millisecond
